+++ core/src/main/java/com/yahoo/ycsb/CommandLine.java
@@ -326,7 +326,7 @@ public class CommandLine
 		  Vector<HashMap<String,ByteIterator>> results=new Vector<HashMap<String,ByteIterator>>();
 		  Status ret=db.scan(table,tokens[1],Integer.parseInt(tokens[2]),fields,results);
 		  System.out.println("Result: "+ret.getName());
-		  int record=0;
+		  long record=0;
 		  if (results.isEmpty())
 		  {
 		     System.out.println("0 records");
